Teletype: page only if screen has more than two rows
The code in Debconf::FrontEnd::Teletype's display_nowrap silently
assumes that the display has more than 2 rows. On a terminal with 2 rows
or less, display_nowrap enters an endless loop, since no actual text
line is ever displayed, only the prompt itself. Since there is no point to try
paging, we completely skip it for 2-or-less-row-terminals.