Dialog: Fix whiptail ellipsized selections
When displaying a select or multiselect using the dialog frontend with
whiptail, if any choices are changed by the ellipsize method (i.e. any
choices are longer than the available screen width to display choices)
they can not be selected.  Running with DEBCONF_DEBUG=developer prints
the message:

    debconf (developer): Input value, "%s" not found in C choices! This should never happen. Perhaps the templates were incorrectly localized.

The cause is that translate_to_C is unable to map the ellipsized choice
text back to the choice value.  This commit adds a hash to do so.  It
also fixes the problem that choices may become ambiguous by allowing
screen overflow in this case and removes a redundant `if defined
$value` conditional.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>