Fix readline prompt for run-parts
When run-parts is executed with the --report flag, it buffers stdout and
stderr from debconf into pipes for formatting. This causes the readline
prompt to be displayed before any of the available options show up, as
readline goes directly to /dev/tty by default and completely sidesteps
the run-parts buffers. The prompt will then block for input, preventing
the options from showing up until the user makes a selection.

This patch forces readline to print to stdout instead of /dev/tty if
both stdout and stderr are being buffered through pipes.

Closes: #928182

Signed-off-by: Heitor Alves de Siqueira <halves@canonical.com>