Fix uninitialized value on "EOF" (^D)
During a `dpkg-reconfigure localepurge`, hitting "EOF" (^D) to one of
the questions will trigger a warning about using an uninitialized value:

```
Use of uninitialized value $_[1] in join or string at /usr/share/perl5/Debconf/DbDriver/Stack.pm line 111.
```

The same happens when hitting "EOF" (^D) when reconfiguring tzdata
(`dpkg-reconfigure --frontend readline tzdata`).

When a user hits ^D on a `readline` or `teletype` front end,
`Debconf::FrontEnd::Teletype->prompt` will return `undef`. The function
description says: "Anything that uses this function should catch that
and handle it, probably by exiting any read/validate loop it is in."

Checking the return value of the `show` function in the `readline`
front end will fix the warning and revert to the backup (if the debconf
is backup capable). The bug should probably fixed in a different
location to also cover the `teletype` front end.

Closes: #723843
LP: #1899343
Signed-off-by: Benjamin Drung <benjamin.drung@canonical.com>